+++ src/components/dialogs/two-windings-transformer/ratio-tap-changer-pane.js
@@ -224,7 +224,7 @@ const RatioTapChangerPane = (props) => {
                 id: rowData.rowIndex,
                 errors: new Map(),
             });
-            const parsedVal = parseInt(newVal);
+            const parsedVal = parseFloat(newVal);
 
             if (!isNaN(parsedVal) && parsedVal >= 0 && parsedVal <= 100) {
                 let tempRows = ratioTapRows;
